Many fixes and tweaks
Lots of fixed things and tweaks to make the offship more functional and flavourful. 

List of changes:

Added wifi variables to mass driver apparatus, so it can fire.
Added a cryogenic oversight console to the cryo-bay, so they actually work.
Adjusted combustion engine injection omni-mixer to be set to take in nothing as an input at round start.
Streamlined power grid design by linking the thruster subgrid line to the EVA subgrid line connecting to the shuttles, rather than running a whole new line under a window outside from the engine room.
Added a breaker box so you can toggle whether the grid is bypassed, rather than it being always bypassed.
Realigned the combustion engine TEG to face the proper direction.
Replaced a stretch of windows towards the back of the combustion chamber with walls to resolve a visual bug.
Added a second atmospherics control console specifically to control the burner mix air injector, rather than both being hooked up to the same injection toggle. This was a really annoying part of the design that just created more busywork for the player.
Removed a shaker cap from the kitchen. The shaker already comes with a cap. 
Removed some excessive grates that didn't need to be there mechanically. They're a useful tool, but there were SO many that it being visually distracting to me.
Made the back wall of hydroponics straight, rather than jagged.
Reorganised medical a little.
Replaced grates under airlocks with appropriate tiles throughout the ship, so that you can click on the tile to easily close the airlock behind you - you can't do this when there's a grate there.
Removed the window behind the new telecommunications server, since it's not much use tucked in the back of a server closet.
Added medkits and some stabilisation kit to medical, and removed some of their blood bags - still more than enough. They shouldn't be very well supplied, this is essentials only.
Set the hydroponics airlock to autoclose. 
Named a bunch of engine things, for clarity, and included a written guide.
Grimed the place up a little.
A bunch more stuff.